\name{makeExampleProtein}
\alias{makeExampleProtein}

\title{Load the coordinates of a protein structure used in many of the package help pages}

\description{This function returns the coordinates of the
Rds3p protein (PDB ID: 2K0A). It is used for the examples in some of the package help pages and can be used 
to test the different functions.}

\usage{
makeExampleProtein()
}


\value{
A list containing a single element, namely a matrix with the 3D coordinates of a the Rds3p protein (109 aminoacids / points)
}


\examples{
\dontrun{
makeExampleProtein()
}
}

\keyword{utilities}
\keyword{proteins}