\name{Rknots-internal}

\alias{edgeIntersections}
\alias{edgeIntersectionsK}
\alias{findComplement}
\alias{bin2N}
\alias{checkVertex}
\alias{xClean}
\alias{buildQuadrilateral}
\alias{rotateQuadrilateral}
\alias{buildTriangles}
\alias{checkTriangles}
\alias{auxiliaryPoints}
\alias{cross}
\alias{triangleIntersection}
\alias{interiorTriangle}
\alias{move3D}
\alias{correspondence}
\alias{contribute}
\alias{msrFast}
\alias{isWholeNumber}
\alias{fillToN}
\alias{lexOrder}
\alias{orientedSign}
\alias{listVertices}
\alias{vertexPresentation}
\alias{findNextE}
\alias{endsAndTriple}
\alias{endComponentCorrection}
\alias{findComponent}
\alias{buildRis2}
\alias{isWholeNumber}
\alias{fillToN}
\alias{lexOrder}
\alias{orientedSign}
\alias{norM}
\alias{listVertices}
\alias{vertexPresentation}
\alias{findNextE}
\alias{endsAndTriple}
\alias{localize}
\alias{findComponent}
\alias{endComponentCorrection}
\alias{singleIntersectionS}
\alias{cp}
\alias{cpr}
\alias{rotFactor}
\alias{escapeDir}
\alias{buildRis}
\alias{escapePath}
\alias{residualIndices}
\alias{computeFactors}
\alias{plotComponent}
\alias{segmentSet}
\alias{cross_product}
\alias{defineCut}
\alias{distancePS}
\alias{distancepstart}
\alias{get2D}
\alias{splitUndercrossing}
\alias{parseToSympy}
\alias{parseToR}
\alias{cyclicP}
\alias{auxiliaryAlexander}

\title{
 Internal Rknots functions
}

\description{
 Internal Rknots functions
}

\details{These are not meant to be called by the user 
(by the way, many of them are low-level functions).
}

\keyword{internal}







